requirements.txt
python-telegram-bot==20.7
sqlalchemy==2.0.23
alembic==1.12.1
python-dotenv==1.0.0
aiohttp==3.9.1
requests==2.31.0
Pillow==10.1.0
APScheduler==3.10.4
pydantic==2.5.2
pydantic-settings==2.1.0
pytz==2023.3
openpyxl==3.1.2
python-docx==1.1.0
reportlab==4.0.7

